requirements.txt
google-generativeai
gTTS